Actualización vistas front

+++ src/main/java/Proyecto_Equipo_7/entidades/Calificacion.java
@@ -28,7 +28,7 @@ public class Calificacion {
     @OneToOne
     private Proveedor proveedor;
 
-    private Integer calificacion;
+    private Double calificacion;
 
     private String comentario;
